requirements.txt
discord.py==1.2.3
SQLAlchemy==1.3.5